delete route owner ship

+++ api/src/tasks/tasks.service.ts
@@ -38,8 +38,8 @@ export class TasksService {
     return this.taskRepository.createTask(createTaskDto, user);
   }
 
-  async deleteTask(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTask(id: number, user: UserEntity): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id }); // deleteはwhereいらんねんね
 
     if (result.affected === 0) {
       throw new NotFoundException(`Task with ID "${id}" not found`);
